edit dto structure
