+++ dto/src/main/java/com/github/punchat/dto/am/access/AccessCodeValidationResult.java
@@ -1,4 +1,4 @@
-package com.github.punchat.dto.am.am.access;
+package com.github.punchat.dto.am.access;
 
 public enum AccessCodeValidationResult {
     VALID, INVALID, OUTDATED
