+++ dto/src/main/java/com/github/punchat/dto/am/access/NewAccessCodeRequest.java
@@ -1,0 +1,12 @@
+package com.github.punchat.dto.am.access;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class NewAccessCodeRequest {
+    private String email;
+}
