+++ dto/src/main/java/com/github/punchat/dto/am/access/WorkspaceAccessCodeValidation.java
@@ -1,4 +1,4 @@
-package com.github.punchat.dto.am.am.invite;
+package com.github.punchat.dto.am.access;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,6 +7,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class WorkspaceEmailValidation {
+public class WorkspaceAccessCodeValidation {
     private String email;
+    private String code;
 }
