+++ dto/src/main/java/com/github/punchat/dto/am/access/WorkspaceAccessCodeValidationResult.java
@@ -1,4 +1,4 @@
-package com.github.punchat.dto.am.am.access;
+package com.github.punchat.dto.am.access;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,7 +7,8 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class WorkspaceAccessCodeValidation {
+public class WorkspaceAccessCodeValidationResult {
     private String email;
     private String code;
+    private AccessCodeValidationResult result;
 }
