--- dto/src/main/java/com/github/punchat/dto/am/am/access/WorkspaceAccessCodeValidationResult.java
@@ -1,15 +1,0 @@
-package com.github.punchat.dto.am.am.access;
-
-import com.github.punchat.dto.am.am.access.AccessCodeValidationResult;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-public class WorkspaceAccessCodeValidationResult {
-    private String email;
-    private String code;
-    private AccessCodeValidationResult result;
-}
