+++ dto/src/main/java/com/github/punchat/dto/am/invite/EmailValidationResult.java
@@ -1,4 +1,4 @@
-package com.github.punchat.dto.am.am.invite;
+package com.github.punchat.dto.am.invite;
 
 public enum EmailValidationResult {
     VALID, INVALID
