--- dto/src/main/java/com/github/punchat/dto/am/messaging/channel/Privacy.java
@@ -1,5 +1,0 @@
-package com.github.punchat.dto.am.messaging.channel;
-
-public enum Privacy {
-    PUBLIC, PRIVATE
-}
