+++ dto/src/main/java/com/github/punchat/dto/messaging/channel/Privacy.java
@@ -1,0 +1,5 @@
+package com.github.punchat.dto.messaging.channel;
+
+public enum Privacy {
+    PUBLIC, PRIVATE
+}
