+++ dto/src/main/java/com/github/punchat/dto/messaging/member/InvitationDto.java
@@ -1,4 +1,5 @@
-package com.github.punchat.dto.am.messaging.user;
+package com.github.punchat.dto.messaging.member;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,6 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class UserDto {
-    private long id;
+public class InvitationDto {
+    private Long userId;
+    private Long role;
 }
