+++ dto/src/main/java/com/github/punchat/dto/messaging/member/MemberDto.java
@@ -1,8 +1,8 @@
-package com.github.punchat.dto.am.messaging.member;
+package com.github.punchat.dto.messaging.member;
 
-import com.github.punchat.dto.am.messaging.role.RoleDto;
-import com.github.punchat.dto.am.messaging.channel.BroadcastChannelDto;
-import com.github.punchat.dto.am.messaging.user.UserDto;
+import com.github.punchat.dto.messaging.role.RoleDto;
+import com.github.punchat.dto.messaging.channel.BroadcastChannelDto;
+import com.github.punchat.dto.messaging.user.UserDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
