+++ dto/src/main/java/com/github/punchat/dto/messaging/resource/ResourceDto.java
@@ -1,5 +1,4 @@
-package com.github.punchat.dto.am.messaging.member;
-
+package com.github.punchat.dto.messaging.resource;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -8,7 +7,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
-public class InvitationDto {
-    private Long userId;
-    private Long role;
+public class ResourceDto {
+    private long id;
 }
