+++ services/am/src/main/java/com/github/punchat/am/domain/access/AccessCodeService.java
@@ -1,7 +1,7 @@
 package com.github.punchat.am.domain.access;
 
-import com.github.punchat.dto.am.am.access.AccessCodeValidationResult;
-import com.github.punchat.dto.am.am.access.WorkspaceAccessCodeValidation;
+import com.github.punchat.dto.am.access.AccessCodeValidationResult;
+import com.github.punchat.dto.am.access.WorkspaceAccessCodeValidation;
 
 public interface AccessCodeService {
     AccessCode generateAccessCode();
