+++ services/am/src/main/java/com/github/punchat/am/domain/invite/workspace/WorkspaceInviteServiceImpl.java
@@ -4,14 +4,14 @@ import com.github.punchat.am.domain.access.AccessCodeService;
 import com.github.punchat.am.domain.invite.InviteService;
 import com.github.punchat.am.domain.invite.State;
 import com.github.punchat.am.events.EventBus;
-import com.github.punchat.dto.am.am.access.AccessCodeValidationResult;
-import com.github.punchat.dto.am.am.access.NewAccessCodeRequest;
-import com.github.punchat.dto.am.am.access.WorkspaceAccessCodeValidation;
-import com.github.punchat.dto.am.am.access.WorkspaceAccessCodeValidationResult;
-import com.github.punchat.dto.am.am.invite.EmailValidationResult;
-import com.github.punchat.dto.am.am.invite.WorkspaceEmailValidation;
-import com.github.punchat.dto.am.am.invite.WorkspaceEmailValidationResult;
-import com.github.punchat.dto.am.am.invite.WorkspaceInvitation;
+import com.github.punchat.dto.am.access.AccessCodeValidationResult;
+import com.github.punchat.dto.am.access.NewAccessCodeRequest;
+import com.github.punchat.dto.am.access.WorkspaceAccessCodeValidation;
+import com.github.punchat.dto.am.access.WorkspaceAccessCodeValidationResult;
+import com.github.punchat.dto.am.invite.EmailValidationResult;
+import com.github.punchat.dto.am.invite.WorkspaceEmailValidation;
+import com.github.punchat.dto.am.invite.WorkspaceEmailValidationResult;
+import com.github.punchat.dto.am.invite.WorkspaceInvitation;
 import com.github.punchat.events.*;
 import com.github.punchat.log.Trace;
 import org.springframework.stereotype.Service;
